Move eq computation to the CPU from the GPU
The GPU has to query the texture 128 times for every pixel. It leads to
huge usage of the GPU. It's way better to do that work once with the CPU
instead.

In the future it could be possible to use a vertex shader for it, but
right now that's weird as it would require changing the entire rendering
pipeline.