Allow a second screen
Here's the way it works in a nutshell:
1. Open a new screen
2. Objects from the visualizations are sent over.
3. Updates are created in animate() for each visualization and sent over
4. In the case of shader, uniforms are applied to the object.

Because they're different contexts and renderers, we have to  a. toJSON and parse the objects as they come over and b. create new textures for those uniforms requiring them.

Another way we could do this is move the entire rendering process to the
second context, but this is a simpler solution and allows rendering on
both (or more?) screens.